Create an imitation_learning directory, called imitation_learning_dir and move generate_bc_trajectories* there.
