+++ compiler_opt/rl/imitation_learning/generate_bc_trajectories.py
@@ -20,7 +20,7 @@ from absl import flags
 from absl import logging
 import gin
 
-from compiler_opt.rl import generate_bc_trajectories_lib
+from compiler_opt.rl.imitation_learning import generate_bc_trajectories_lib
 from compiler_opt.tools import generate_test_model  # pylint:disable=unused-import
 
 from tf_agents.system import system_multiprocessing as multiprocessing
